[MOD] Speed up Python 0014
Read the forums and got the idea to implement a hash of all past
sequence numbers. Added this to vastly speed up the times and also
remove the i % 2 check in the loop... which I had no idea what it
was there for, and there is no proof to suggest that the answer
could not have come from an even number.